+++ src/lib/Core/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Updated density data model.
 * Removed the support for the DSN6 density data format.
 * Fixed a bug in marching cubes that sometimes added bad triangles because of a vertex index cache not being properly cleaned.
+* Marching cubes annotation tweak.
 
 # 3.0.4
 * Fix in PDB CRYST1 record parsing.